Create nWaysCoinChange
How many different ways can you make change for an amount, given a list of coins? In this problem, your code will need to efficiently compute the answer.

Reference Video : https://www.youtube.com/watch?v=_fgjrs570YE&list=PLrmLmBdmIlpsHaNTPP_jHHDx_os9ItYXr&index=10

if(j>=coins(i))
    T[i][j] = T[i-1,j]+T[i, j-coins(i)]
else
    T[i][j] = T[i-1][j]
Explanation Missing in Video:
Suppose T[2][5] = 5. It is calculated by T[1][5] + T [2][2]. It is saying that there are 5 different ways to have total value of 5 by these 3 different coins. 
The idea from the first part is that assume we don't have the current coin available (3), then the number of ways we can have from the other 2s are 3 (T[1][5] = 3, i.e.: 11111, 2111,221). 
The second part is to deduct the total from the current coin's value (here 3) and add the remaining values's way to the total number of ways. You know why? Because what it means is how many ways we can find for "total - coin[i]", here 5 - 3 = 2 by all of the coins. It means, 2 can be formed by 11 or 2 which later we can add another 3 for the deducted valued to have the total value: 311, 32.